+++ packages/web3-providers/src/Etherscan/apis/RedPacketAPI.ts
@@ -6,13 +6,13 @@ import type { RedPacketBaseAPI } from '../../entry-types.js'
 import { fetchJSON } from '../../entry-helpers.js'
 
 export class EtherscanRedPacketAPI implements RedPacketBaseAPI.Provider<ChainId, SchemaType> {
-    async getHistories(
+    async getHistoryTransactions(
         chainId: ChainId,
         senderAddress: string,
         contractAddress: string,
         methodId: string,
-        startBlock?: number,
-        endBlock?: number,
+        startBlock: number,
+        endBlock: number,
     ): Promise<Array<Transaction<ChainId, SchemaType>> | undefined> {
         if (!senderAddress || !contractAddress || !startBlock || !endBlock || !methodId) return
