+++ packages/web3-providers/src/types/RedPacket.ts
@@ -1,17 +1,27 @@
 import type { Pageable, PageIndicator } from '@masknet/shared-base'
+import type { RedPacketJSONPayloadFromChain } from '@masknet/web3-shared-evm'
 import type { Transaction, NonFungibleCollection } from '@masknet/web3-shared-base'
 import type { HubOptions_Base } from '../entry-types.js'
 
 export namespace RedPacketBaseAPI {
     export interface Provider<ChainId, SchemaType> {
-        getHistories(
+        getHistories?: (
             chainId: ChainId,
             senderAddress: string,
             contractAddress: string,
             methodId: string,
-            startBlock?: number,
-            endBlock?: number,
-        ): Promise<Array<Transaction<ChainId, SchemaType>> | undefined>
+            startBlock: number,
+            endBlock: number,
+        ) => Promise<RedPacketJSONPayloadFromChain[] | undefined>
+
+        getHistoryTransactions?: (
+            chainId: ChainId,
+            senderAddress: string,
+            contractAddress: string,
+            methodId: string,
+            startBlock: number,
+            endBlock: number,
+        ) => Promise<Array<Transaction<ChainId, SchemaType>> | undefined>
 
         /** Get non-fungible collections owned by the given account. */
         getCollectionsByOwner?: (
